Fix: Jenkins slave recreated from image taken of micro instance doesn't use ephemeral storage for build directory
We now always create /mnt/ephemeral (or whatever the mount point for ephemeral storage is) and symlink to it. This works independently whether the ephemeral volume is mounted or not . And it allows using images from ephemeral-less instances types to be used on instance types that have ephemeral storage, and vice versa.