Eliminated spark-node /etc/hosts entry; cosmetics
Spark does a reverse lookup of the value of SPARK_LOCAL_IP (or the default) if that value is an IP. The reverse lookup appears to consider /etc/hosts. If the value (or the default) is the box' default IP, the resolution will yield the hostname "spark-node". The driver then advertises itself to executors using an Akka URL that contains that hostname, e.g.

akka.tcp://sparkDriver@spark-node:56712

The problem is that on an executor this hostname resolves to the host the executor is running on. IOW, the executor will try to reach the driver on its own host, not the host the driver is actually running on. Eliminating the spark-node entry will fix this. A driver running on the master will advertise itself as

akka.tcp://sparkDriver@spark-master:56712

A driver running on a slave will advertise itself using the configured hostname:

akka.tcp://sparkDriver@ip-172-31-21-133.us-west-2.compute.internal:56712