Downgrade Toil Jenkins worker to Spark 1.5.2 (resolves #189)
Downgrades the Toil Jenkins test worker to Spark 1.5.2, which is the same
version as the Docker image that we use for Apache Spark. If the two versions
are different, then the Toil unittests may not be able to connect to the
Spark cluster we've spawned. Specifically, the driver<->leader wire protocol is
subject to change between different Spark major versions.

It turned out that the SPARK_HOME environment variable was already set, so
I did not need to change anything there.